requirements.txt
Django>=4.2
djangorestframework
scikit-learn
pandas
joblib
python-dotenv
